feat: tag translator 변경
